Fixed NoopSpan and most of gitignore
