+++ .gitignore
@@ -2,3 +2,5 @@
 .gradle/
 **/.idea/workspace.xml
 **/.idea/tasks.xml
+.metadata/**
+**/.project
